Refactored less driver focused methods into seperate class, improved driver shutdown
