+++ omroneipApp/Db/testReadNoPacking.db
@@ -1,40 +1,87 @@
-#Packing does not effect the behaviour of this record.
-record(waveform, "$(P)readUDTNoPacking") {
+# Here we send and recieve 11 seperate messages
+
+record(waveform, "$(P)readUDTPackingElement1") {
     field(SCAN, "I/O Intr")
     field(FTVL, "UCHAR")
     field(DTYP, "asynInt8ArrayIn")
-    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[1] UDT 9 0 &allow_packing=0")
+    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[1] UDT 1 none &allow_packing=0")
     field(NELM, "1996")
 }
 
-record(waveform, "$(P)readUDTNoPackingElement1") {
+record(waveform, "$(P)readUDTPackingElement2") {
     field(SCAN, "I/O Intr")
     field(FTVL, "UCHAR")
     field(DTYP, "asynInt8ArrayIn")
-    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[1] UDT 1 0 &allow_packing=0")
+    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[2] UDT 1 none &allow_packing=0")
     field(NELM, "1996")
 }
 
-record(waveform, "$(P)readUDTNoPackingElement2") {
+record(waveform, "$(P)readUDTPackingElement3") {
     field(SCAN, "I/O Intr")
     field(FTVL, "UCHAR")
     field(DTYP, "asynInt8ArrayIn")
-    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[2] UDT 1 0 &allow_packing=0")
+    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[3] UDT 1 none &allow_packing=0")
     field(NELM, "1996")
 }
 
-record(waveform, "$(P)readUDTNoPackingElement3") {
+record(waveform, "$(P)readUDTPackingElement4") {
     field(SCAN, "I/O Intr")
     field(FTVL, "UCHAR")
     field(DTYP, "asynInt8ArrayIn")
-    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[3] UDT 1 0 &allow_packing=0")
+    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[4] UDT 1 none &allow_packing=0")
     field(NELM, "1996")
 }
 
-record(waveform, "$(P)readUDTNoPackingElement4") {
+record(waveform, "$(P)readUDTPackingElement5") {
     field(SCAN, "I/O Intr")
     field(FTVL, "UCHAR")
     field(DTYP, "asynInt8ArrayIn")
-    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[4] UDT 1 0 &allow_packing=0")
+    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[5] UDT 1 none &allow_packing=0")
     field(NELM, "1996")
+}
+
+record(waveform, "$(P)readUDTPackingElement6") {
+    field(SCAN, "I/O Intr")
+    field(FTVL, "UCHAR")
+    field(DTYP, "asynInt8ArrayIn")
+    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[6] UDT 1 none &allow_packing=0")
+    field(NELM, "1996")
+}
+
+record(waveform, "$(P)readUDTPackingElement7") {
+    field(SCAN, "I/O Intr")
+    field(FTVL, "UCHAR")
+    field(DTYP, "asynInt8ArrayIn")
+    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[7] UDT 1 none &allow_packing=0")
+    field(NELM, "1996")
+}
+
+record(waveform, "$(P)readUDTPackingElement8") {
+    field(SCAN, "I/O Intr")
+    field(FTVL, "UCHAR")
+    field(DTYP, "asynInt8ArrayIn")
+    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[8] UDT 1 none &allow_packing=0")
+    field(NELM, "1996")
+}
+
+record(waveform, "$(P)readUDTPackingElement9") {
+    field(SCAN, "I/O Intr")
+    field(FTVL, "UCHAR")
+    field(DTYP, "asynInt8ArrayIn")
+    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) ISISTestMultiple[9] UDT 1 none &allow_packing=0")
+    field(NELM, "1996")
+}
+
+record(waveform, "$(P)testChildArray") {
+    field(SCAN, "I/O Intr")
+    field(FTVL, "UCHAR")
+    field(DTYP, "asynInt8ArrayIn")
+    field(INP, "@asyn($(PORT), 0, 2)@$(POLLER) testChildArray[1] UDT 2 none &allow_packing=0")
+    field(NELM, "1996")
+}
+
+record(longin, "$(P)readUInt32") {
+    field(SCAN, "I/O Intr")
+    field(DTYP, "asynInt32")
+    field(INP, "@asyn($(PORT), 0, 1)@$(POLLER) testUDINT UDINT 1 none &allow_packing=0")
 }
