+++ omroneipApp/src/drvOmroneip.h
@@ -44,6 +44,7 @@
 
 #define CREATE_TAG_TIMEOUT 1000 //ms
 
+
 typedef enum {
   dataTypeBool,
   dataTypeInt,
@@ -63,35 +64,133 @@ typedef enum {
   MAX_OMRON_DATA_TYPES
 } omronDataType_t;
 
+typedef struct
+{
+  omronDataType_t dataType;
+  const char *dataTypeString;
+} omronDataTypeStruct;
+
+// Supported PLC datatypes and the strings which users use to read/write this type
+static omronDataTypeStruct omronDataTypes[MAX_OMRON_DATA_TYPES] = {
+    {dataTypeBool, "BOOL"},
+    {dataTypeInt, "INT"},
+    {dataTypeDInt, "DINT"},
+    {dataTypeLInt, "LINT"},
+    {dataTypeUInt, "UINT"},
+    {dataTypeUDInt, "UDINT"},
+    {dataTypeULInt, "ULINT"},
+    {dataTypeReal, "REAL"},
+    {dataTypeLReal, "LREAL"},
+    {dataTypeString, "STRING"},
+    {dataTypeWord, "WORD"},
+    {dataTypeDWord, "DWORD"},
+    {dataTypeLWord, "LWORD"},
+    {dataTypeUDT, "UDT"},
+    {dataTypeUDT, "TIME"}};
+
+// This stores information about each communication tag to the PLC.
+// A new instance will be made for each record which requsts to uniquely read/write to the PLC
+struct omronDrvUser_t
+{
+  size_t startIndex;
+  size_t sliceSize;
+  std::string tag;
+  std::string dataType;
+  int32_t tagIndex;
+  size_t dataCounter;
+  std::string pollerName;
+  size_t tagOffset;
+  double timeout;
+  std::string optimisationFlag;
+  size_t strCapacity;
+  bool readFlag;
+};
+
 typedef std::unordered_map<std::string, std::vector<std::string>> structDtypeMap;
 typedef std::unordered_map<std::string, std::string> drvInfoMap;
 
 struct omronDrvUser_t;
 class omronEIPPoller;
+class omronUtilities;
 
 /* Main class for the driver */
 class epicsShareClass drvOmronEIP : public asynPortDriver {
 public:
-  drvOmronEIP(const char *portName,
-              char *gateway,
-              char *path,
-              char *plcType,
-              int debugLevel);
-  ~drvOmronEIP();
+   const char *driverName = "drvOmronEIP"; /* String for asynPrint */
+   drvOmronEIP(const char *portName,
+               char *gateway,
+               char *path,
+               char *plcType,
+               int debugLevel);
+   ~drvOmronEIP();
+
 
    bool omronExiting_;
+
    /* All reading of data is initiated from this function which runs at a predefined frequency */
    void readPoller();
+   /* Creates a new instance of the omronEIPPoller class and starts a new thread named after this new poller which reads data linked to the poller name.*/
+   asynStatus createPoller(const char * portName, const char * pollerName, double updateRate);
+   /* Reimplemented from asynDriver. This is called when each record is loaded into epics. It processes the drvInfo from the record and attempts
+      to create a libplctag tag and an asynParameter. It saves the handles to these key objects within the tagMap_. This tagMap_ is then used to
+      process read and write requests to the driver.*/
+   asynStatus drvUserCreate(asynUser *pasynUser, const char *drvInfo, const char **pptypeName, size_t *psize)override;
+   /* Improves efficiency by merging duplicate tags and looking for situations where multiple UDT field read requests can be replaced with a single UDT read */
+   asynStatus optimiseTags();
    /* Takes a csv style file, where each line contains a structure name followed by a list of datatypes within the struct
-      Stores the user input struct as a map containing Struct:field_list pairs. It then calls createStructMap and passes this map */   
+   Stores the user input struct as a map containing Struct:field_list pairs. It then calls createStructMap and passes this map */   
    asynStatus loadStructFile(const char * portName, const char * filePath);
+
+   /* The read interface only needs to be reimplemented for int8Arrays as the other read interfaces have helper functions to set the value of the
+      asynParameters that they write to */
+   asynStatus readInt8Array(asynUser *pasynUser, epicsInt8 *value, size_t nElements, size_t *nIn)override;
+
+   /* Write interfaces reimplemented from asynPortDriver. Write data from the asynParameter to the associated libplctag tag*/
+   asynStatus writeInt8Array(asynUser *pasynUser, epicsInt8 *value, size_t nElements)override;
+   asynStatus writeUInt32Digital(asynUser *pasynUser, epicsUInt32 value, epicsUInt32 mask)override;
+   asynStatus writeInt32(asynUser *pasynUser, epicsInt32 value)override;
+   asynStatus writeInt64(asynUser *pasynUser, epicsInt64 value)override;
+   asynStatus writeFloat64(asynUser *pasynUser, epicsFloat64 value)override;
+   asynStatus writeOctet(asynUser *pasynUser, const char * value, size_t nChars, size_t* nActual)override;
+
+private:
+   bool initialized_; // Tracks if the driver successfully initialized
+   bool startPollers_;
+   std::string tagConnectionString_; // Stores the basic PLC connection information common to all read/write requests
+   // The key is the name of the struct, the vector is a list of byte offsets within the structure
+   std::unordered_map<std::string, std::vector<int>> structMap_;
+   /* The key is the name of the struct, the vector contains strings representing the dtypes and embbed structs/arrays. 
+   Used to match user requests to offsets in the structMap */
+   structDtypeMap structDtypeMap_;
+   structDtypeMap structRawMap_;
+   std::unordered_map<std::string, omronEIPPoller*> pollerList_ = {}; // Stores the name of each registered poller
+   /* Maps the index of each registered asynParameter to essential communications data for the parameter */
+   std::unordered_map<int, omronDrvUser_t*> tagMap_;
+   omronUtilities *utilities; //smart pointerify
+   friend class omronUtilities;
+};
+
+/* Class which stores information about each poller */
+class omronEIPPoller{
+  public:
+      omronEIPPoller(const char* portName, const char* pollerName, double updateRate);
+      ~omronEIPPoller();
+      const char* belongsTo_;
+      const char* pollerName_;
+      double updateRate_;
+};
+
+/* Class which contains generic functions required by the driver */
+class omronUtilities {
+public:
+   drvOmronEIP *pDriver;
+   asynUser *pasynUserSelf;
+   const char * driverName;
+   omronUtilities(drvOmronEIP *pDriver);
+   ~omronUtilities();
+
    /* Loops through each structure within the map and calls findOffsets which creates the final structure offset map */
    asynStatus createStructMap(structDtypeMap rawMap);
-   /* Takes the elements of a user defined structure requested by the user in the drvInfo string and finds their offsets
-      from the structMap_. The main purpose of this algorithm is to skip over embedded structs/arrays as we count through the map.
-      For example, if I had a structure A with 2 elements, but the first was a 100byte structure, referencing A[1] would need to skip over this
-      100 byte structure. */
-   int findRequestedOffset(std::vector<size_t> indices, std::string structMap);
    /* Recursively extract datatypes from an embedded array, this array may contain embedded structs in which case expandStructsRecursive is called */
    std::vector<std::string> expandArrayRecursive(structDtypeMap const& rawMap, std::string arrayDesc);
    /* Recursively extract datatypes from embedded structures and list them in the correct order */
@@ -109,53 +208,13 @@ public:
    std::string findArrayDtype(structDtypeMap const& expandedMap, std::string arrayDesc);
    /* Returns the largest standard datatype within a structure, this includes embedded structures and arrays and is used for calculating padding*/
    int getBiggestDtype(structDtypeMap const&, std::string structName);
-   /* Creates a new instance of the omronEIPPoller class and starts a new thread named after this new poller which reads data linked to the poller name.*/
-   asynStatus createPoller(const char * portName, const char * pollerName, double updateRate);
-   /* Reimplemented from asynDriver. This is called when each record is loaded into epics. It processes the drvInfo from the record and attempts
-      to create a libplctag tag and an asynParameter. It saves the handles to these key objects within the tagMap_. This tagMap_ is then used to
-      process read and write requests to the driver.*/
-   asynStatus drvUserCreate(asynUser *pasynUser, const char *drvInfo, const char **pptypeName, size_t *psize)override;
+   /* Takes the elements of a user defined structure requested by the user in the drvInfo string and finds their offsets
+      from the structMap_. The main purpose of this algorithm is to skip over embedded structs/arrays as we count through the map.
+      For example, if I had a structure A with 2 elements, but the first was a 100byte structure, referencing A[1] would need to skip over this
+      100 byte structure. */
+   int findRequestedOffset(std::vector<size_t> indices, std::string structMap);
    /* This is responsible for parsing drvInfo when records are created. It takes the drvInfo string and parses it for required data.
       It returns a map of all of the data required by the driver to setup the asyn parameter and a boolean which indicates the validity of the data. 
       This function is called twice, once before and once after database initialization, the first time we just return asynDisabled. */
    drvInfoMap drvInfoParser(const char *drvInfo);
-   /* Improves efficiency by merging duplicate tags and looking for situations where multiple UDT field read requests can be replaced with a single UDT read */
-   asynStatus optimiseTags();
-   
-   /* The read interface only needs to be reimplemented for int8Arrays as the other read interfaces have helper functions to set the value of the
-      asynParameters that they write to */
-   asynStatus readInt8Array(asynUser *pasynUser, epicsInt8 *value, size_t nElements, size_t *nIn)override;
-
-   /* Write interfaces reimplemented from asynPortDriver. Write data from the asynParameter to the associated libplctag tag*/
-   asynStatus writeInt8Array(asynUser *pasynUser, epicsInt8 *value, size_t nElements)override;
-   asynStatus writeUInt32Digital(asynUser *pasynUser, epicsUInt32 value, epicsUInt32 mask)override;
-   asynStatus writeInt32(asynUser *pasynUser, epicsInt32 value)override;
-   asynStatus writeInt64(asynUser *pasynUser, epicsInt64 value)override;
-   asynStatus writeFloat64(asynUser *pasynUser, epicsFloat64 value)override;
-   asynStatus writeOctet(asynUser *pasynUser, const char * value, size_t nChars, size_t* nActual)override;
-
-   private:
-   bool initialized_; // Tracks if the driver successfully initialized
-   bool startPollers_;
-   std::string tagConnectionString_; // Stores the basic PLC connection information common to all read/write requests
-   // The key is the name of the struct, the vector is a list of byte offsets within the structure
-   std::unordered_map<std::string, std::vector<int>> structMap_;
-   /* The key is the name of the struct, the vector contains strings representing the dtypes and embbed structs/arrays. 
-   Used to match user requests to offsets in the structMap */
-   structDtypeMap structDtypeMap_;
-   structDtypeMap structRawMap_;
-   std::unordered_map<std::string, omronEIPPoller*> pollerList_ = {}; // Stores the name of each registered poller
-   std::unordered_map<int, omronDrvUser_t*> tagMap_;
-   /* Maps the index of each registered asynParameter to essential communications data for the parameter */
-};
-
-/* Class which stores information about each poller */
-class omronEIPPoller{
-  public:
-      omronEIPPoller(const char* portName, const char* pollerName, double updateRate);
-      ~omronEIPPoller();
-      const char* belongsTo_;
-      const char* pollerName_;
-      double updateRate_;
-
 };
